slight refactor for synths
